fix : 특정 콘서트 상세 정보 조회 API 호출 위치 수정
